requirements.txt
autopep8
django
djangorestframework
django-rest-auth
django-allauth
django-cors-headers
